requirement.txt
django
pymysql
django-ckeditor
pillow
mysqlclient
django-jsonfield-backport